+++ src/Ext/Rules/Body.h
@@ -288,6 +288,8 @@ public:
 		Valueable<bool> NoRearm_Temporal;
 		Valueable<bool> NoReload_UnderEMP;
 		Valueable<bool> NoReload_Temporal;
+		Valueable<double> VeteranReload;
+		Nullable<double> VeteranEmptyReload;
 		Valueable<bool> NoTurret_TrackTarget;
 
 		Valueable<bool> GatherWhenMCVDeploy;
@@ -350,6 +352,7 @@ public:
 		Valueable<int> WarheadAnimZAdjust;
 
 		Valueable<bool> IvanBombAttachToCenter;
+		Valueable<bool> MissileSpawnAttackCell;
 
 		Valueable<bool> FallingDownTargetingFix;
 		Valueable<bool> AIAirTargetingFix;
@@ -526,6 +529,11 @@ public:
 		Nullable<int> PoseDir_Field;
 
 		Valueable<bool> ApproachTarget_StopWhenInRange;
+		Valueable<bool> ReadyToNextMission_MovingCheck;
+
+		Valueable<bool> NoAlphaImageOnBuildup;
+
+		Valueable<bool> Warhead_PreventScatter;
 
 		ExtData(RulesClass* OwnerObject) : Extension<RulesClass>(OwnerObject)
 			, Storage_TiberiumIndex { -1 }
@@ -767,6 +775,8 @@ public:
 			, NoRearm_Temporal { false }
 			, NoReload_UnderEMP { false }
 			, NoReload_Temporal { false }
+			, VeteranReload { 1.0 }
+			, VeteranEmptyReload {}
 			, NoTurret_TrackTarget { false }
 			, GatherWhenMCVDeploy { true }
 			, AIFireSale { true }
@@ -818,6 +828,7 @@ public:
 			, WarheadAnimZAdjust { -15 }
 
 			, IvanBombAttachToCenter { false }
+			, MissileSpawnAttackCell { true }
 
 			, FallingDownTargetingFix { false }
 			, AIAirTargetingFix { false }
@@ -988,6 +999,10 @@ public:
 			, PoseDir_Field{}
 
 			, ApproachTarget_StopWhenInRange { false }
+
+			, NoAlphaImageOnBuildup { false }
+			, ReadyToNextMission_MovingCheck { false }
+			, Warhead_PreventScatter { false }
 		{ }
 
 		virtual ~ExtData() = default;